disable or enable legend
I'm suggesting this modification to omit legend because I'm working with a large number of species and the built-in legend of the function doesn't show all of them. I'm manually creating a separate legend to identify the species but I didn't find a way to omit the default legend of the plot. I think this modification would benefit others with the same problem.